Ability to define options on a per element basis.
The option optionsForElement should specify a function that is passed
the current element and returns option overrides to be used while
counting the value of the current element.  This is useful when the
jQuery selector matches multiple input or textarea elements and the
counter element or counting behavior must be specific to each selected
element.